# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SIPD API Documentation",
        "description": "Dokumentasi API untuk aplikasi SIPD",
        "contact": {
            "email": "admin@sipd_new.test"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://sipd_new.test/api/",
            "description": "Local Server"
        }
    ],
    "paths": {
...
        "/projects": {
            "get": {
                "tags": [
                    "Project"
                ],
                "operationId": "07f276c6e629646d087ee4b728e1234c",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Response Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "meta": {
                                            "properties": {
                                                "status": {
                                                    "type": "integer",
                                                    "example": 200
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "success show all project"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": "1"
                                                    },
                                                    "code_project": {
                                                        "type": "string",
                                                        "example": "PRJ-20260827-WL0R"
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "Kampung Nelayan Merah Putih"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "Pengajuan izin analisis dampak pembangunan program Kampung Nelayan Merah Putih."
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "pagination": {
                                            "properties": {
                                                "total": {
                                                    "type": "integer",
                                                    "example": 20
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 4
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Project"
                ],
                "operationId": "d91af95183d5934b42aec1325d5165db",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProjectRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Response Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "meta": {
                                            "properties": {
                                                "status": {
                                                    "type": "integer",
                                                    "example": 201
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "create project success"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": "1"
                                                },
                                                "code_project": {
                                                    "type": "string",
                                                    "example": "PRJ-20260827-WL0R"
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "Kampung Nelayan Merah Putih"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "Pengajuan izin analisis dampak pembangunan program Kampung Nelayan Merah Putih."
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/projects/{id}": {
            "put": {
                "tags": [
                    "Project"
                ],
                "operationId": "546bd69cc1e45c7e4a50c3da131d0969",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProjectRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Response Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "meta": {
                                            "properties": {
                                                "status": {
                                                    "type": "integer",
                                                    "example": 201
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "success update project"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": "1"
                                                },
                                                "code_project": {
                                                    "type": "string",
                                                    "example": "PRJ-20260827-WL0R"
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "Kampung Nelayan Merah Putih"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "Pengajuan izin analisis dampak pembangunan program Kampung Nelayan Merah Putih."
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Project"
                ],
                "operationId": "650fd22a9e728f1de77f375a5ec4e414",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Response Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "meta": {
                                            "properties": {
                                                "status": {
                                                    "type": "integer",
                                                    "example": 201
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "success remove project"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": null,
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "ProjectRequest": {
                "title": "Project Request",
                "description": "",
                "properties": {
                    "title": {
                        "type": "string",
                        "example": "Kampung Nelayan Merah Putih"
                    },
                    "description": {
                        "type": "string",
                        "example": "Pengajuan izin analisis dampak pembangunan Kampung Nelayan Merah Putih."
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "Sanctum Token"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}